Tracking GPS_TRACK, Split up files to gps_handler
